+++ js/main.js
@@ -26,6 +26,7 @@ async function start() {
 
   // Zanim cokolwiek narysujemy, ustalamy skąd biorą się dane.
   await DATA.init();
+  await wczytajSpisOkladek();
 
   renderAccount();
   applyAuthVisibility();